Create ImageToNumpyFile
Turns image files into .npy files. This program was used to change the way we store the data. Before we had a file containing a folder with training and validation data, which contained folders with 10 different thickness containing all the crystals. Now all the data is put into 3 files, one contains the images, another contains the thickness (classification) and the last one is a text file containing a key which can be used to see which crystal number corresponds to which exact crystal.