Update and rename SymTest to DataManipulation/SymmetryTest
Added Comments. This data checks to see if the full sizes images are symmetric horizontally and vertically. This is done by comparing the 1st quadrant with the other 2nd, 3rd and 4th quadrant by calculating the rms average. A value of 0 indicated a perfect symmetry large values indicate a difference.